# reduce fillInStackTrace expenses due to ArgNormalizer