--- a/common/scala/src/main/scala/whisk/core/entity/FullyQualifiedEntityName.scala
+++ b/common/scala/src/main/scala/whisk/core/entity/FullyQualifiedEntityName.scala
@@ -48,7 +48,7 @@ protected[core] case class FullyQualifiedEntityName(path: EntityPath, name: Enti
    */
   def add(n: EntityName) = FullyQualifiedEntityName(path.addPath(name), n)
 
-  def toDocId = DocId(qualifiedName)
+  def toDocId = new DocId(qualifiedName)
   def namespace: EntityName = path.root
   def qualifiedNameWithLeadingSlash: String = EntityPath.PATHSEP + qualifiedName
   def asString = path.addPath(name) + version.map("@" + _.toString).getOrElse("")