ArgNormalizer uses a try/catch pattern when parsing entity names. in two common cases, the arg to be normalized is known to be a string. we can fast-path this case. i called this `DocId.direct()`

the expense of this can be witnessed from call stack samples: you will see `fillInStackTrace` due to exceptions being thrown by spray's json parser.

preliminary experiments show that this results in more stable performance, i.e. with improvements principally to p99 latency numbers.

```
           p25   p50   p75    p90    p99     max
master     20ms  23ms  28ms   40ms   270ms  1.7s
this PR    18ms  21ms  26ms   35ms   85ms   1.4s
```

test protocol: using wrk -t8 -c800; wipe.yml, then warmup 60s, then timed run 120s
endpoint: blocking invokes of echo action
deployment: environments/local, local couchdb, 44-way with 48GB memory; 12 invokers
load driver: 16-way located in the same rack

PG2/2715  🔵