@starpit 2018-01-27
for future reference, here is an example call stack for one of the two call sites:

```
	at java.lang.Throwable.fillInStackTrace(Native Method)
	at java.lang.Throwable.fillInStackTrace(Throwable.java:783)
	- locked <0x00000000e2187818> (a spray.json.JsonParser$ParsingException)
	at java.lang.Throwable.<init>(Throwable.java:265)
	at java.lang.Exception.<init>(Exception.java:66)
	at java.lang.RuntimeException.<init>(RuntimeException.java:62)
	at spray.json.JsonParser$ParsingException.<init>(JsonParser.scala:31)
	at spray.json.JsonParser.fail(JsonParser.scala:213)
	at spray.json.JsonParser.value(JsonParser.scala:64)
	at spray.json.JsonParser.parseJsValue(JsonParser.scala:43)
	at spray.json.JsonParser$.apply(JsonParser.scala:28)
	at spray.json.PimpedString.parseJson(package.scala:45)
	at whisk.core.entity.ArgNormalizer$$anonfun$factory$1.apply(ArgNormalizer.scala:30)
	at whisk.core.entity.ArgNormalizer$$anonfun$factory$1.apply(ArgNormalizer.scala:30)
	at scala.util.Try$.apply(Try.scala:192)
	at whisk.core.entity.ArgNormalizer$class.factory(ArgNormalizer.scala:30)
	at whisk.core.entity.DocId$.factory(DocInfo.scala:107)
	at whisk.core.entity.ArgNormalizer$class.apply(ArgNormalizer.scala:45)
	at whisk.core.entity.DocId$.apply(DocInfo.scala:107)
	at whisk.core.entity.FullyQualifiedEntityName.toDocId(FullyQualifiedEntityName.scala:51)
```