@markusthoemmes 2018-01-27
Woa. That actually looks like loads of overhead generated without us knowing it. @rabbah could we possibly just ditch `ArgNormalizer`? We shouldn't have any `null` values anyway and I'm not aware of any values where we'd need to trim?