@dubeejw 2018-01-27
Quick reference of everything that uses ArgNormalizer: https://github.com/apache/incubator-openwhisk/search?p=1&q=%22ArgNormalizer%22&type=&utf8=%E2%9C%93.

Double whammy here? `val activationId = Try { ActivationId(n) }`
https://github.com/apache/incubator-openwhisk/blob/20a5a6083d1630037740b4cf62f0d4e8fbc5c829/core/controller/src/main/scala/whisk/core/controller/Activations.scala#L95