@starpit 2018-01-27
made the constructor public, and got rid of `DocId.direct`. i didn't change any other call sites (they don't matter, for performance, yet at least)