@starpit 2018-01-28
@dubeejw this is a problem, though indirectly. i've opened a separate issue for it: #3228

i'll work out a solution tomorrow. it should offer a nice improvement, if my investigations are on target.